documentation

+++ database/src/secondary.rs
@@ -20,7 +20,7 @@ pub struct Entry {
 pub fn read(
     buffer: &mut BytesMut,
     secondary_file: &File,
-    chunk_number: u32,
+    chunk_number: u64,
 ) -> io::Result<Box<[BlockInfo]>> {
     let secondary_size = secondary_file.metadata()?.len() as usize;
     read_buf(secondary_file, buffer, 0, secondary_size)?;
